New - yum::remove definition and related refactoring
It was desired to make the package absence of the yum-presto and
PackageKit-yum-plugin packages part of yum::core where they seemed to be
a better fit.  This was largely desired because the very initial runs of
puppet might try to use the delta-rpm support provided by yum-presto.
Since yum::core is a dependency of all Package resources, it was thought
that this would be a good way to accomplish that.

However, that set up a circular dependency, so I went to change those to
Package resources to Exec wrappers around 'yum remove' and then realized
that the code base was becoming littered with these and that a new
definition was in order, hence yum::remove.

This contains all of that, plus the transition of other existing
Exec['yum remove'] instances to use the new yum::remove definition.

I *think* this is better, but it's still fugly.  Nonetheless, it should
make repairs easier should a better solution come along.