New - rewrite of Class[yum::cron]
This class no longer accepts a source URI for the configuration content.
Instead it uses a template and expects the consumer to provide any
non-default values through parameters which now cover most every
configurable aspect.  The ones remaining at their defaults which are not
controllable by parameters are rarely changed, but could certainly be
parameterized in the future.